edits

+++ src/elisa/binary_system/surface/mesh.py
@@ -408,8 +408,8 @@ def _generate_neck_zs(delta_z, component, neck_position, neck_polynomial):
         curve = np.column_stack((x_curve, z_curve))
         lengths = up.sqrt(np.sum(np.diff(curve, axis=0) ** 2, axis=1))
         neck_lengths = np.cumsum(lengths)
-        num_z = neck_lengths[-1] // delta_z
-        segments = np.linspace(0, neck_lengths[-1], num=int(num_z))[:-1]
+        num_z = int(neck_lengths[-1] // delta_z)
+        segments = np.linspace(0, neck_lengths[-1], num=num_z)[:-1]
 
         z_ns = np.interp(segments, neck_lengths, x_curve[:-1])
         r_neck = np.polyval(neck_polynomial, z_ns)
@@ -445,7 +445,7 @@ def trapezoidal_overcontact_neck_points(
     phi, z, z_ns, r_neck, separator = _generate_neck_zs(delta_z, component, neck_position, neck_polynomial)
 
     for ii, zz in enumerate(z_ns):
-        num = const.HALF_PI * r_neck[ii] // delta_z
+        num = int(const.HALF_PI * r_neck[ii] // delta_z)
         num = num + 1 if num < 5 else num
         phis = np.linspace(0, const.HALF_PI, num=int(num), endpoint=False)[1:]
         z_n = np.full(phis.shape, zz)
